avsync: send VIDEO_STOP to tsync [1/1]
PD#TV-31230

At the beginning of playback, audio HAL tries to update apts and find
the gap between audio and video is too big, as a result sync mode is set to
VMASTER. When next AUDIO_START comes, the start pts is not set to PCR
because of the wrong mode. The video PTS in use here is from last
playback.

Send VIDEO_STOP will clean the last video PTS. It will prevent mode
switch logic.

Change-Id: Ic8f08d95975366fee42b137af8c51f6cfc7c1691