avsynclib: handle pts and vsync alignment [1/1]
PD#SWPL-58495

Problem:
Stream is 23.98 fps, output is 59.94 fps. So each frame need to be hold
to two and a half VSYNC. The initial phase is set to half VSYNC. So the
frame PTS will be too close to VSYNC every 5 VSYNCs.

Solution:
Detect this case and adjust the phase to separate PTS from VSYNC before
pattern is detected.

Verify:
AH221 + DV cert file in JIRA

Change-Id: Ic3377cae78f305d65377b60d97fc11aa6d620615